+++ faux_macros/src/methods.rs
@@ -15,8 +15,10 @@ pub struct Args {
 pub struct Mockable {
     // the real definitions inside the impl block
     real: syn::ItemImpl,
-    // the morphed definitions and the _where_ mehods
-    morphed_and_wheres: syn::ItemImpl,
+    // the morphed definitions
+    morphed: syn::ItemImpl,
+    // the _when_ methods in their own impl
+    whens: syn::ItemImpl,
     // path to real struct
     real_ty: syn::TypePath,
     // path to morphed struct
@@ -49,30 +51,73 @@ impl Mockable {
         // start transforming
         let real_ty = real_ty(&morphed_ty, args.path);
 
-        let mut morphed_and_wheres = real.clone();
+        let mut morphed = real.clone();
 
-        let mut methods = morphed_and_wheres
-            .items
-            .iter_mut()
-            .filter_map(|item| match item {
-                syn::ImplItem::Method(m) => Some(m),
-                _ => None,
-            });
+        let mut methods = morphed.items.iter_mut().filter_map(|item| match item {
+            syn::ImplItem::Method(m) => Some(m),
+            _ => None,
+        });
 
         let mut when_methods = vec![];
         for func in &mut methods {
-            let signature = Signature::morph(&mut func.sig, &func.vis);
+            let signature = Signature::morph(
+                &mut func.sig,
+                real.trait_.as_ref().map(|(_, path, _)| path),
+                &func.vis,
+            );
             func.block = signature.create_body(args.self_type, &real_ty, &morphed_ty)?;
             if let Some(when_method) = signature.create_when() {
                 when_methods.push(syn::ImplItem::Method(when_method));
             }
         }
 
-        morphed_and_wheres.items.extend(when_methods);
+        let generics = match &morphed_ty.path.segments.last().unwrap().arguments {
+            syn::PathArguments::AngleBracketed(generics_in_struct) => {
+                let generics_in_struct = &generics_in_struct.args;
+                let params: syn::punctuated::Punctuated<_, _> = real
+                    .generics
+                    .params
+                    .iter()
+                    .filter(|generic| match generic {
+                        syn::GenericParam::Type(ty) => generics_in_struct.iter().any(|g| match g {
+                            syn::GenericArgument::Type(syn::Type::Path(type_path)) => {
+                                type_path.path.is_ident(&ty.ident)
+                            }
+                            _ => false,
+                        }),
+                        syn::GenericParam::Lifetime(lt) => {
+                            generics_in_struct.iter().any(|g| match g {
+                                syn::GenericArgument::Lifetime(lifetime_in_struct) => {
+                                    lt.lifetime == *lifetime_in_struct
+                                }
+                                _ => false,
+                            })
+                        }
+                        syn::GenericParam::Const(_) => true,
+                    })
+                    .cloned()
+                    .collect();
+                syn::Generics {
+                    params,
+                    ..real.generics.clone()
+                }
+            }
+            _ => syn::Generics::default(),
+        };
+
+        let (impl_generics, _, where_clause) = generics.split_for_impl();
+
+        let whens = syn::parse2(quote! {
+            impl #impl_generics #morphed_ty #where_clause {
+                #(#when_methods) *
+            }
+        })
+        .unwrap();
 
         Ok(Mockable {
             real,
-            morphed_and_wheres,
+            morphed,
+            whens,
             real_ty,
             morphed_ty,
         })
@@ -83,7 +128,8 @@ impl From<Mockable> for proc_macro::TokenStream {
     fn from(mockable: Mockable) -> Self {
         let Mockable {
             real,
-            morphed_and_wheres,
+            morphed,
+            whens,
             real_ty,
             morphed_ty,
         } = mockable;
@@ -93,7 +139,14 @@ impl From<Mockable> for proc_macro::TokenStream {
         let mod_ident = {
             let ident = &real_ty.path.segments.last().unwrap().ident;
             syn::Ident::new(
-                &format!("_faux_real_impl_{}", ident),
+                &match &real.trait_ {
+                    None => format!("_faux_real_impl_{}", ident),
+                    Some((_, trait_, _)) => format!(
+                        "_faux_real_impl_{}_{}",
+                        ident,
+                        trait_.segments.last().unwrap().ident
+                    ),
+                },
                 proc_macro2::Span::call_site(),
             )
         };
@@ -101,7 +154,9 @@ impl From<Mockable> for proc_macro::TokenStream {
         // make the original methods at least pub(super)
         // since they will be hidden in a nested mod
         let mut real = real;
-        publicize_methods(&mut real);
+        if !real.trait_.is_some() {
+            publicize_methods(&mut real);
+        }
         let real = real;
 
         // creates an alias `type Foo = path::to::RealFoo` that may be wrapped inside some mods
@@ -122,9 +177,16 @@ impl From<Mockable> for proc_macro::TokenStream {
 
             // type Foo = path::to::RealFoo
             let alias = {
+                // the alias has to be public up to the mod wrapping it
+                let pub_supers = if path_to_ty.len() < 2 {
+                    quote! {}
+                } else {
+                    let supers = std::iter::repeat(quote! { super }).take(path_to_ty.len() - 1);
+                    quote! { pub(#(#supers)::*) }
+                };
                 let pathless_type = path_to_ty.pop().unwrap();
                 quote! {
-                    pub type #pathless_type = #path_to_real_from_alias_mod;
+                    #pub_supers type #pathless_type = #path_to_real_from_alias_mod;
                 }
             };
 
@@ -135,7 +197,9 @@ impl From<Mockable> for proc_macro::TokenStream {
         };
 
         proc_macro::TokenStream::from(quote! {
-            #morphed_and_wheres
+            #morphed
+
+            #whens
 
             mod #mod_ident {
                 // make everything that was in-scope above also in-scope in this mod